backend/requirements.txt
# ET_model Backend Dependencies
# Install with: pip install -r backend/requirements.txt

# Core ML
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
xgboost>=2.0
joblib>=1.3

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0

# Authentication
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-dotenv>=1.0

# External API
requests>=2.28
